Refactor MyRecipes to Container/View Components
